Rename view submodules to be lowercase and reorganize timeline subpkg
