+++ sift/view/cameras.py
