+++ sift/view/texture_atlas.py
