+++ sift/view/timeline/scene.py
@@ -25,8 +25,8 @@ from PyQt4.QtGui import QGraphicsScene, QPen, QBrush, QPainter, QGraphicsView, Q
 from PyQt4.QtOpenGL import QGLFormat, QGL, QGLWidget
 
 from sift.common import flags, span
-from sift.view.TimelineCommon import VisualState, CoordTransform
-from sift.view.TimelineItems import QTrackItem, QFrameItem, QTimeRulerItem
+from .common import VisualState, CoordTransform
+from .items import QTrackItem, QFrameItem, QTimeRulerItem
 
 LOG = logging.getLogger(__name__)
